streamlit move slow REST pull to subpage
